fix the trip location deletion issue in deleting trip plan

+++ backend/services/operations/tripPlanDeleteOperations.js
@@ -1,6 +1,7 @@
 const mongoose = require("mongoose");
 
 const tripPlanController = require("./../controllers/tripPlanController.js");
+const tripLocationController = require("./../controllers/tripLocationController.js");
 const partnerLocationController = require("./../controllers/partnerLocationController.js");
 const itemBoughtController = require("./../controllers/itemBoughtController.js");
 
@@ -14,11 +15,19 @@ module.exports.deleteTripPlan = (tripPlanId) => {
     await session.withTransaction(async () => {
       const tripPlan = await tripPlanController.findById(tripPlanId).lean().session(session);
 
+      if (!tripPlan) {
+        const error = new Error(`Trip Plan with ID=${tripPlanId} not found!`); error.code = 404;
+        throw error;
+      }
+
       // Delete trip locations from partner locations
       await partnerLocationController.deleteAssociatedTripLocationsFromPartnerLocations(tripPlan.tripLocations, session);
 
       // Delete items boughts
       const itemBoughtsDeleted = await itemBoughtController.deleteItemBoughtsByTripLocations(tripPlan.tripLocations, session);
+
+      // Delete the trip locations completely
+      await tripLocationController.deleteByIds(tripPlan.tripLocations, session);
 
       // Delete the trip plan
       await tripPlanController.deleteById(tripPlanId, session);
